LoginBackend/src/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==1.10.13
email-validator==1.3.1